The last merge from developMaster

+++ app/controllers/documents_controller.rb
@@ -5,46 +5,47 @@ class DocumentsController < AuthenticatedController
   @after_new_user_created = ""
 
   def index
-    @documents = Document.for_user(current_user)
+    @documents = policy_scope(Document).each { |d| authorize d }
     session[:ret_url] = "/documents"
   end
 
-  def show; end
+  def show
+    authorize @document
+  end
 
   def new
-    @document = Document.new(base_params.slice(:category, :group))
-    documents_helper = DocumentService.new(:category => @document.category);
-    @cards = documents_helper.get_card_values(current_user)
+    @document = Document.new(base_params.slice(:category, :group).merge(user: current_user))
+
+    authorize @document
+
+    @cards = card_values(@document.category)
   end
 
   def edit
+    authorize @document
+
     session[:ret_url] = get_return_url_path
     @shares = @document.shares
   end
 
   def create
-    @document = Document.new(document_params.merge(:user_id => current_user.id))
+    options = document_params.merge(user: current_user)
+    @document = Document.new(options)
+
+    authorize @document
+
     respond_to do |format|
-      if @document.save && @document.update(document_share_params) #TODO: dynamic route builder for categories
-        if is_new_contact_creating
-          save_return_url_path(@document.id)
-          format.html { redirect_to new_contact_path :redirect => @after_new_user_created }
-        end
-        if return_url?
-          format.html { redirect_to session[:ret_url], notice: 'Document was successfully created.' }
-        else
-          format.html { redirect_to documents_path, notice: 'Document was successfully created.' }
-        end
-        format.json { render :show, status: :created, location: @document }
+      if @document.save && @document.update(document_share_params)
+        handle_document_saved(format)
       else
-        @cards = DocumentService.new(:category => @document.category).get_card_values(current_user)
-        format.html { render :new }
-        format.json { render json: @document.errors, status: :unprocessable_entity }
+        handle_document_not_saved(format)
       end
     end
   end
 
   def update
+    authorize @document
+
     respond_to do |format|
       if is_new_contact_creating
         save_return_url_path(params[:id])
@@ -68,6 +69,8 @@ class DocumentsController < AuthenticatedController
   end
 
   def destroy
+    authorize @document
+
     @document.destroy
     S3Service.delete_from_storage(@document.url)
     redirect_page = session[:ret_url] || documents_path
@@ -78,22 +81,29 @@ class DocumentsController < AuthenticatedController
   end
   
   def get_drop_down_options
-    category = params[:category]
-    documents_helper = DocumentService.new(:category => params[:category], :user => current_user);
-    render :json => documents_helper.get_card_values(current_user).flatten
+    render :json => card_values(params[:category]).flatten
   end
 
   private
 
+  def card_values(category)
+    service = DocumentService.new(:category => category)
+    service.get_card_values(resource_owner)
+  end
+
+  def resource_owner
+    @document.present? ? @document.user : current_user
+  end
+
   def set_contacts
-    @contacts = Contact.for_user(current_user)
-    @contacts_shareable = @contacts.reject { |c| c.emailaddress == current_user.email } 
+    contact_service = ContactService.new(:user => resource_owner)
+    @contacts = contact_service.contacts
+    @contacts_shareable = contact_service.contacts_shareable
   end
 
   def set_document
-    @document = Document.for_user(current_user).find(params[:id])
-    documents_helper = DocumentService.new(:category => @document.category);
-    @cards = documents_helper.get_card_values(current_user)
+    @document = Document.find(params[:id])
+    @cards = card_values(@document.category)
   end
 
   def base_params
@@ -101,11 +111,11 @@ class DocumentsController < AuthenticatedController
   end
   
   def document_share_params
-    share_service = ShareService.new(user_id: current_user.id, contact_ids: params[:document][:contact_ids])
+    share_service = ShareService.new(user_id: resource_owner.id, contact_ids: params[:document][:contact_ids])
     share = share_service.fill_document_share
     #cleare document shares before updating current document
     share_service.clear_shares(@document)
-    document_params.merge(:shares_attributes => share, :user_id => current_user.id)
+    document_params.merge(:shares_attributes => share, :user_id => resource_owner.id)
   end
 
   def document_params
@@ -146,5 +156,25 @@ class DocumentsController < AuthenticatedController
   
   def set_document_update_date_to_now(document)
     document.updated_at = Time.now.utc
+  end
+
+  def handle_document_saved(format)
+    #TODO: dynamic route builder for categories
+    if is_new_contact_creating
+      save_return_url_path(@document.id)
+      format.html { redirect_to new_contact_path :redirect => @after_new_user_created }
+    end
+    if return_url?
+      format.html { redirect_to session[:ret_url], notice: 'Document was successfully created.' }
+    else
+      format.html { redirect_to documents_path, notice: 'Document was successfully created.' }
+    end
+    format.json { render :show, status: :created, location: @document }
+  end
+
+  def handle_document_not_saved(format)
+    @cards = card_values(@document.category)
+    format.html { render :new }
+    format.json { render json: @document.errors, status: :unprocessable_entity }
   end
 end
